pyproject.toml
[build-system]
requires = ["setuptools<80", "wheel", "versioneer[toml]"]
build-backend = "setuptools.build_meta"

[project]
name = "ms-mint-app"
description = "Metabolomics Integrator Interactive Web-Application (Mint)"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
  {name = "Mario E. Valdes-Tresanco"},
  {name = "Mario S. Valdes-Tresanco"},
  {name = "Soren Wacker"},
]
maintainers = [
  {name = "Mario E. Valdes-Tresanco"},
  {name = "Mario S. Valdes-Tresanco"},
  {name = "Soren Wacker"},
]
keywords = ["metabolomics", "integrator", "mint"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent"
]
dependencies = [
  "asari-metabolomics",
  "dash==3.3.0",
  "dash-bootstrap-components",
  "dash[diskcache]==3.3.0",
  "dash_extensions",
  "diskcache",
  "distinctipy==1.3.4",
  "duckdb==1.4.1",
  "fastcluster",
  "fastparquet",
  "feffery-antd-components==0.4.2",
  "feffery-utils-components==0.3.3",
  "flask-caching",
  "flask-sqlalchemy",
  "flask_login",
  "lxml>=6.0.2",
  "lttbc",
  "matplotlib>=3.10.8",
  "numpy==2.3.3",
  "openpyxl",
  "pandas==2.3.2",
  "plotly==6.3.1",
  "polars-lts-cpu",
  "polars==1.33.1",
  "psutil",
  "pyarrow==18.0.0",
  "pygixml",
  "PyYAML",
  "pyteomics",
  "scikit-learn",
  "scipy>=1.16.3",
  "seaborn",
  "sqlalchemy",
  "tqdm>=4.67.1",
  "waitress",
  "werkzeug==3.1.5",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/LewisResearchGroup/ms-mint-app"
Documentation = "https://lewisresearchgroup.github.io/ms-mint-app/"
Repository = "https://github.com/LewisResearchGroup/ms-mint-app.git"
"Bug Tracker" = "https://github.com/LewisResearchGroup/ms-mint-app/issues"

[project.scripts]
Mint = "ms_mint_app.scripts.Mint:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ms_mint_app*"]

[tool.setuptools.package-data]
"ms_mint_app" = ["assets/*", "static/*", "static/**/*"]

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "src/ms_mint_app/_version.py"
tag_prefix = "v"
parentdir_prefix = ""

[tool.setuptools.dynamic]
version = { attr = "ms_mint_app.__version__" }
